abstracting communication manager
